feat: implement the interaction of views with prepaired business scenarios
