+++ server/apps/accounts/implemented/__init__.py
@@ -1,0 +1,1 @@
+from .signup import signup_implementation # Noqa
