--- server/apps/accounts/views/__init__.py
@@ -1,3 +1,0 @@
-from .login_user import login_user  # Noqa: F401
-from .logout_user import logout_user  # Noqa: F401
-from .register_user import register_user  # Noqa: F401
