--- server/apps/accounts/views/common_constants.py
@@ -1,7 +1,0 @@
-from typing import Final
-
-
-LOGIN_SUCCESS_URL_NAME: Final = 'main:profile'
-LOGIN_PAGE_TEMPLATE_NAME: Final = 'accounts/login.html'
-REGISTER_SUCCESS_URL_NAME: Final = 'accounts:login'
-REGISTRATION_PAGE_TEMPLATE_NAME: Final = 'accounts/register.html'
